+++ app/src/androidTest/java/com/grammatek/simaromur/UtteranceCacheManagerTest.java
@@ -22,6 +22,7 @@ import com.grammatek.simaromur.cache.Utterance;
 import com.grammatek.simaromur.cache.UtteranceCache;
 import com.grammatek.simaromur.cache.UtteranceCacheManager;
 import com.grammatek.simaromur.cache.VoiceAudioDescription;
+import com.grammatek.simaromur.utils.FileUtils;
 
 import java.io.File;
 import java.io.IOException;
