+++ app/src/main/java/com/grammatek/simaromur/utils/FileUtils.java
@@ -1,4 +1,4 @@
-package com.grammatek.simaromur;
+package com.grammatek.simaromur.utils;
 
 import static java.nio.file.Files.deleteIfExists;
